scripts/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyYAML==6.0.1
lyricsgenius==3.0.1
